Moved reject_member method from BaseView into ProjectView
